requirements.txt
Flask==3.1.1
python-docx
num2words